% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_theme_torpeda.R
\name{colors_discrete_torpeda}
\alias{colors_discrete_torpeda}
\alias{colors_diverging_torpeda}
\title{toRpEDA color palettes for ggplot2 objects}
\usage{
colors_discrete_torpeda(n = 2)

colors_diverging_torpeda()
}
\arguments{
\item{n}{number of colors for color palette}
}
\value{
color palette as vector of charactes
}
\description{
toRpEDA color palettes for ggplot2 objects
}
