% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Correlation matrix}
\usage{
cor_matrix(data, variables = NULL, method = "pearson")
}
\arguments{
\item{data}{A numeric matrix, a data frame or a vector.}

\item{variables}{An optional character string giving a variables for computing correlation matrix. This must be colnames from data.}

\item{method}{A character string specifying the correlation coefficient to be calculated. Possible values are: 'pearson' - computes the Pearson correlation coefficient (default), 'spearman' - computes the Spearman rank correlation coefficient.}
}
\value{
A correlation matrix is returned as a numeric matrix, which is a symmetric matrix with correlation coefficients between -1 and 1. The diagonal elements are always equal to 1.
}
\description{
'cor_matrix' computing correlation matrix for numeric variables from data.
}
\examples{

library(toRpEDA)

#computing correlation matrix only from numeric variables
cor_matrix(data = iris)

#computing correlation matrix from selected numeric variables
cor_matrix(data = mtcars, variables = c("cyl", "vs", "carb"))

#computing correlation matrix from numeric variables using 'spearman' method
data <- data.frame(x = rnorm(100), y = rnorm(100), z = rnorm(100))
cor_matrix(data, method = 'spearman')


}
