% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_values.R
\name{find_missing_values}
\alias{find_missing_values}
\title{The aim of this function is to check how many missing values has our data.}
\usage{
find_missing_values(df, variable = NULL)
}
\arguments{
\item{df}{dataframe}

\item{variable}{selected column (default value = all columns)}
}
\value{
A table and a diagram representing the number of missing values in each/selected colums.
}
\description{
The aim of this function is to check how many missing values has our data.
}
\examples{
library(ggplot2)
df <- data.frame(data.frame(a = c(2,5,4,NA, 5,7,NA), b =  c(NA,NA,2,NA, 5,7,3),
c = c(23,5,4,1, 5,7,NA), d =  c(NA,NA,2,NA, 5,7,5),
e = c(4,5,4,NA, 0,7,NA), f =  c(1,123,2,87, 5,7,6),
g = c(253,NA,NA,NA, NA,NA,NA), h =  c(3,34,2,5, 5,7,56)))

find_missing_values(df)
find_missing_values(df, variable = c("a", "e"))
find_missing_values(df, variable = "b")

}
