% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stat.R
\name{get_descriptive_stat}
\alias{get_descriptive_stat}
\title{Generates descriptive statistics for a given data frame.
Continuous variables stats are grouped in a single data frame while each
categorical variable has its stats stored in a separate data frame.}
\usage{
get_descriptive_stat(df, vars = NULL, include_long_catg = FALSE)
}
\arguments{
\item{df}{A data.frame object to analyse}

\item{vars}{Variables to compute stats for. NULL means all variables}

\item{include_long_catg}{Whether or not to include stats for categorical
variables with more than 50 unique values}
}
\value{
a list of data frames with descriptive statistics
}
\description{
Generates descriptive statistics for a given data frame.
Continuous variables stats are grouped in a single data frame while each
categorical variable has its stats stored in a separate data frame.
}
\examples{
get_descriptive_stat(iris)
get_descriptive_stat(iris, vars = c('Petal.Length', 'Species'))
get_descriptive_stat(iris, include_long_catg = TRUE)
}
