% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_data.R
\name{impute_mean}
\alias{impute_mean}
\title{Impute mean}
\usage{
impute_mean(df, variables = colnames(df))
}
\arguments{
\item{df}{A data frame containing the data to impute.}

\item{variables}{A character vector containing the names of the columns to impute
using mean. The default value is a list containing all column names of `df`.}
}
\value{
A data frame with the same dimensions as `df` where missing values have been imputed using the mean.
}
\description{
This function imputes mean
The function can impute mean for all columns or for a selected set of columns.
}
\examples{
df <- data.frame(x <- c(1, NA, 3, NA),y <- c(NA, 4, 6, 8))
impute_mean(df)

}
