% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_data.R
\name{impute_missing_data}
\alias{impute_missing_data}
\title{Impute missing data in a data frame}
\usage{
impute_missing_data(df, variables = list(colnames(df)), methods = c("mean"))
}
\arguments{
\item{df}{A data frame containing the data to impute.}

\item{variables}{A list of character vectors specifying the columns to impute for each method.
Each element of the list should be a character vector containing the names of the columns to impute
using the corresponding method in the `methods` parameter. The default value is a list containing all column names of `df`.}

\item{methods}{A character vector specifying the methods to use for imputing missing data.
The available methods are `"mean"`, `"median"` and `"mode"`. The default value is `"mean"`.}
}
\value{
A data frame with the same dimensions as `df` where missing values have been imputed using the specified methods.
}
\description{
This function imputes missing data in a data frame using one or more specified methods.
The function can impute missing data for all columns or for a selected set of columns.
}
\examples{
df <- data.frame(x <- c(1, NA, 3, NA),y <- c(NA, 4, 6, 8))
impute_missing_data(df)

}
