% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Function for a given data frame or chosen variables either scales them to
interval [0,1], or transforms them by subtracting mean and dividing by
standard deviation.}
\usage{
normalise(
  x,
  variables = colnames(x[, which(sapply(x, is.numeric))]),
  mode = "standard",
  na.om = FALSE
)
}
\arguments{
\item{x}{data frame}

\item{variables}{vector of names of variables to consider; all numeric
variables are taken by default.}

\item{mode}{either "standard" (the deafult) to perform standardisation or
"interval" for scaling variables to unit interval.}

\item{na.om}{logical. Should whole row be removed when missing value
is present?}
}
\value{
Data frame inputed with normalised chosen columns.
}
\description{
Function for a given data frame or chosen variables either scales them to
interval [0,1], or transforms them by subtracting mean and dividing by
standard deviation.
}
\examples{
library('toRpEDA')

# Standardising variables of data frame
normalise(iris)

# scaling variables of data frame to unit integral
normalise(iris, mode = "interval")

# Choosing variables for normalisation
normalise(iris, variables = "Sepal.Width")

}
