% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{This function returns principal components and their associated eigenvectors - PCA}
\usage{
PCA(df, standardize = TRUE, plot_explained_variance = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{standardize}{- boolean - whether you want to automatically standardize your data}

\item{plot_explained_variance}{- boolean - whether you want to plot the proportion of explained variance by next vectors}
}
\value{
The function returns eigenvectors with their according principal components in PCA. Additionally displays a plot of explained variance in data.
}
\description{
This function returns principal components and their associeted eigenvectors - PCA
}
\examples{
library(ggplot2)
PCA(mtcars, standardize = TRUE, plot_explained_variance = TRUE)
}
