% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_plot.R
\name{plot_bar_plot}
\alias{plot_bar_plot}
\title{Bar plots by target}
\usage{
plot_bar_plot(df, target, variables = NULL, col_number = 2)
}
\arguments{
\item{df}{data frame}

\item{target}{character, name of target column}

\item{variables}{vector of column names for which charts are to be made}

\item{col_number}{numeric the number of columns in which the charts will be displayed}
}
\value{
None
}
\description{
Creates a plot of categorical columns by target
}
\examples{
my_data <- data.frame(
      target = c("R", "Python", "R", "R", "Python", "Python"),
      a = c(1, 2, 3, 4, 2, 2),
      b = c("a", "b", "c", "b", "c", "c"),
      c = c(TRUE, FALSE, TRUE, FALSE, FALSE, TRUE),
      d = c("dog ", "cat", "dog", "dog", "dog", "cat"),
      e = c("PL", "PL", "PL", "USA", "USA", "USA"),
      f = c("Europa", "Europa", "America", "Europa", "America", "Europa")
      )
plot_bar_plot(my_data, "target")
plot_bar_plot(my_data, "target", variables = c("b", "d", "e"))
plot_bar_plot(my_data, "target", col_number = 3)
}
