% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_qual.R
\name{plot_bar_qual}
\alias{plot_bar_qual}
\title{This function returns bar plots of quality columns of a dataframe}
\usage{
plot_bar_qual(df, variables = NULL)
}
\arguments{
\item{df}{A data frame}

\item{variables}{A vector of names (or numbers) of columns of which the user wants to create bar plots}
}
\value{
Bar plots in k rows and length(variables)/k columns where k is the largest
divisor of the number of columns fewer than the number of chosen columns
}
\description{
This function returns bar plots of quality columns of a dataframe
}
\examples{
plot_bar_qual(iris, 5)
}
