% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_vars.R
\name{plot_hist_vars}
\alias{plot_hist_vars}
\title{plot_hist_vars}
\usage{
plot_hist_vars(data, vars, plots_per_page = 6)
}
\arguments{
\item{data}{a data frame}

\item{vars}{a character vector of variable names to include in the plots.
If NULL (default), all numeric variables are used.}

\item{plots_per_page}{an integer indicating the number of plots to display per
page. Default is 6.}
}
\value{
series of histogram plots to the R console.
}
\description{
The plot_hist_vars function creates a series of histogram plots for
the numeric variables of a data frame. The function allows users to
specify which variables to include in the plots and how many plots to display
per page. The function makes use of the ggplot2 and patchwork packages to
create the plots and combine them into a single page.
}
\examples{
library(ggplot2)
library(patchwork)

data(mtcars)
plot_hist_vars(mtcars, c("mpg", "cyl"))

data(mtcars)
plot_hist_vars(mtcars)

data(iris)
plot_hist_vars(iris, c("Sepal.Length", "Petal.Length", "Species"))

data(iris)
plot_hist_vars(iris)
}
