% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_num_plots.R
\name{plot_num_plots}
\alias{plot_num_plots}
\title{Draws numeric plots (boxplots and violin plots) by target
for selected columns.}
\usage{
plot_num_plots(df, variables = colnames(df), target, plot_type = "boxplot")
}
\arguments{
\item{df}{data frame with numeric values}

\item{variables}{column names to be included}

\item{target}{name of target column}

\item{plot_type}{type of plots, 'boxplot' or 'violin'}
}
\value{
list of plots
}
\description{
Draws numeric plots (boxplots and violin plots) by target
for selected columns.
}
\examples{
library("toRpEDA")

# drawing 4 plots
plot_num_plots(iris, target = "Species", plot_type="violin")
plot_num_plots(iris, target = "Species", plot_type="boxplot")

plot_num_plots(mtcars, target = "gear", plot_type="boxplot")
plot_num_plots(mtcars, variables = c("wt", "mpg"), target = "gear", plot_type="boxplot")

}
