% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{QQ Plot}
\usage{
qqplot(data, y = NULL, variables = NULL)
}
\arguments{
\item{data}{A matrix, data frame or numeric vector, which is sample for qqplot.}

\item{y}{An optional vector specifying the values of a custom distribution to compare against the dataset's quantiles. If provided, the QQ plot will be generated using this custom distribution instead of the standard normal distribution.}

\item{variables}{An optional character string specifying the variables of dataset to include in the QQ plot. If not provided, all numeric variables from the dataset will be used.}
}
\description{
The qqplot function generates a QQ plot for a given dataset or a vector. By default, it plots the QQ plot against the standard normal distribution. However, if a vector y is provided, it will plot the QQ plot against a different specified distribution.
}
\examples{

library(toRpEDA)

qqplot(iris)

# QQ plot with selected variables
qqplot(mtcars, variables = c("cyl", "carb", "mpg"))

# QQ plot with custom distribution
custom_distribution <- rt(150, df=4)
qqplot(iris$Sepal.Length, y = custom_distribution)

}
