% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport_data_info.R
\name{raport_data_info}
\alias{raport_data_info}
\title{Generate report_data_info}
\usage{
raport_data_info(
  df,
  variables = NULL,
  output_file = NULL,
  output_dir = getwd()
)
}
\arguments{
\item{df}{A data frame.}

\item{variables}{A vector includes names of columns .}

\item{output_file}{A string giving the name of the output file. Default is NULL.}

\item{output_dir}{A string giving the directory where the output file will be saved. Default is the current working directory.}
}
\value{
Report generated to the local file. It contains information about variables types, data dimensions,
missing values, potential ID/redundant columns, outliers and mismatched value types
metrics for
}
\description{
Generate report_data_info
}
\examples{
\dontrun{
report_data_info(iris, output_file = "iris_report.html")
}


}
