% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one.R
\name{table_one}
\alias{table_one}
\title{Dataset Information
Function generates a list of the most common statistics
such as mean, standard deviation, minimum and maximum
grouped by selected argument}
\usage{
table_one(numeric_vars, groupby, data)
}
\arguments{
\item{numeric_vars}{column names of numeric columns of dataframe which we want to use to have table one
arguments should be typed as character vector}

\item{groupby}{element which we want to group by the rest numeric variables}

\item{data}{Data frame}
}
\value{
returns a list of the most commonly used statistical measures
such as mean, standard deviation, minimum and maximum
of desired numerical variables
}
\description{
Dataset Information
Function generates a list of the most common statistics
such as mean, standard deviation, minimum and maximum
grouped by selected argument
}
\examples{
library("toRpEDA")
library(dplyr)
library(stats)
table_one(numeric_vars = c("Petal.Length","Petal.Width"), groupby = "Species", data= iris)

}
