% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Function that finds potential outlier observations}
\usage{
outliers(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
An information about outliers
You can learn more about package authoring with RStudio at:
}
\description{
Function that finds potential outlier observations
}
\examples{
library(isotree)
library(toRpEDA)
library(glue)

outliers(iris)

a <- c(1,2,3,42,1,1,0)
outliers(a)
}
