% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundant_cols.R
\name{redundant_cols}
\alias{redundant_cols}
\title{redundant_cols`()` suggests redundant columns.}
\usage{
redundant_cols(
  df,
  variables = colnames(df),
  correlated = FALSE,
  delete = FALSE
)
}
\arguments{
\item{df}{Data frame.}

\item{variables}{Vector of names of variables to consider. All of variables
are taken as default.}

\item{correlated}{If `TRUE`, function treats highly correlated variables as redundant columns and also deletes
them if `delete` is `TRUE`. `FALSE` as default.}

\item{delete}{If `TRUE`, function returns data frame with deleted redundant
columns. `FALSE` as default.}
}
\value{
If `delete` = `FALSE`, vector of names of redundant columns. Otherwise
data frame with deleted redundant columns. In case there are not any redundant
columnm, `character(0)` is returned.
}
\description{
redundant_cols`()` suggests redundant columns.
}
\examples{
library("toRpEDA")
redundant_cols(iris)
redundant_cols(iris, c("Sepal.Length","Sepal.Width","Species"), delete = TRUE)

}
