% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one(1).R
\name{table_one}
\alias{table_one}
\title{Dataset Information
Function generates a data frame of the most common statistics
such as mean, standard deviation, minimum and maximum
grouped by selected argument for numerical variables}
\usage{
table_one(numeric_vars = NULL, groupby, data)
}
\arguments{
\item{numeric_vars}{column names of numeric columns of dataframe which we want to use to have table one
arguments should be typed as character vector}

\item{groupby}{element which we want to group by the rest numeric variables but only one grupby element}

\item{data}{Data frame}
}
\value{
returns a list of the most commonly used statistical measures
such as mean, standard deviation, minimum and maximum
of desired numerical variables
}
\description{
Dataset Information
Function generates a data frame of the most common statistics
such as mean, standard deviation, minimum and maximum
grouped by selected argument for numerical variables
}
\examples{
library("toRpEDA")
library(dplyr)
library(stats)
table_one(numeric_vars = c("Petal.Length","Petal.Width"), groupby = "Species", data= iris)
table_one(groupby = 'Species', data= iris)
}
